+++ docs/changelog.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-^
+
+### Added
+
+* parameter `regularize` to `make`.
+
 ### Changed
 
 * Project was moved to the [Rayference](https://github.com/rayference) organization.